requirements.txt
 streamlit 
 sumy
 PyPDF2
 nltk
 python-docx
